build: remove //chrome/browser/icon_loader source copies
