+++ atom/browser/atom_browser_main_parts.cc
@@ -21,6 +21,7 @@
 #include "base/command_line.h"
 #include "base/threading/thread_task_runner_handle.h"
 #include "chrome/browser/browser_process.h"
+#include "chrome/browser/icon_manager.h"
 #include "chrome/browser/net/chrome_net_log_helper.h"
 #include "components/net_log/chrome_net_log.h"
 #include "components/net_log/net_export_file_writer.h"
@@ -307,6 +308,12 @@ AtomBrowserMainParts::GetGeolocationControl() {
       content::ServiceManagerConnection::GetForProcess()->GetConnector();
   connector->BindInterface(device::mojom::kServiceName, std::move(request));
   return geolocation_control_.get();
+}
+
+IconManager* AtomBrowserMainParts::GetIconManager() {
+  if (!icon_manager_.get())
+    icon_manager_.reset(new IconManager);
+  return icon_manager_.get();
 }
 
 }  // namespace atom
