+++ chromium_src/chrome/browser/browser_process.cc
@@ -4,15 +4,13 @@
 
 #include "chrome/browser/browser_process.h"
 
-#include "chrome/browser/icon_manager.h"
 #include "chrome/browser/printing/print_job_manager.h"
 #include "ui/base/l10n/l10n_util.h"
 
 BrowserProcess* g_browser_process = NULL;
 
 BrowserProcess::BrowserProcess()
-    : print_job_manager_(new printing::PrintJobManager),
-      icon_manager_(new IconManager) {
+    : print_job_manager_(new printing::PrintJobManager) {
   g_browser_process = this;
 }
 
@@ -26,12 +24,6 @@ void BrowserProcess::SetApplicationLocale(const std::string& locale) {
 
 std::string BrowserProcess::GetApplicationLocale() {
   return locale_;
-}
-
-IconManager* BrowserProcess::GetIconManager() {
-  if (!icon_manager_.get())
-    icon_manager_.reset(new IconManager);
-  return icon_manager_.get();
 }
 
 printing::PrintJobManager* BrowserProcess::print_job_manager() {
