+++ chromium_src/chrome/browser/browser_process.h
@@ -15,8 +15,6 @@
 
 #include "base/macros.h"
 
-class IconManager;
-
 namespace printing {
 class PrintJobManager;
 }
@@ -30,13 +28,11 @@ class BrowserProcess {
 
   void SetApplicationLocale(const std::string& locale);
   std::string GetApplicationLocale();
-  IconManager* GetIconManager();
 
   printing::PrintJobManager* print_job_manager();
 
  private:
   std::unique_ptr<printing::PrintJobManager> print_job_manager_;
-  std::unique_ptr<IconManager> icon_manager_;
   std::string locale_;
 
   DISALLOW_COPY_AND_ASSIGN(BrowserProcess);
