+++ filenames.gni
@@ -580,12 +580,6 @@ filenames = {
     "chromium_src/chrome/browser/browser_process.h",
     "chromium_src/chrome/browser/chrome_process_finder_win.cc",
     "chromium_src/chrome/browser/chrome_process_finder_win.h",
-    "chromium_src/chrome/browser/icon_loader_mac.mm",
-    "chromium_src/chrome/browser/icon_loader_win.cc",
-    "chromium_src/chrome/browser/icon_loader.cc",
-    "chromium_src/chrome/browser/icon_loader.h",
-    "chromium_src/chrome/browser/icon_manager.cc",
-    "chromium_src/chrome/browser/icon_manager.h",
     "chromium_src/chrome/browser/chrome_notification_types.h",
     "chromium_src/chrome/browser/extensions/global_shortcut_listener.cc",
     "chromium_src/chrome/browser/extensions/global_shortcut_listener.h",
@@ -652,7 +646,6 @@ filenames = {
     "chromium_src/library_loaders/libspeechd.h",
   ]
 
-  lib_sources_linux = [ "chromium_src/chrome/browser/icon_loader_auralinux.cc" ]
   lib_sources_nss = [
     "chromium_src/chrome/browser/certificate_manager_model.cc",
     "chromium_src/chrome/browser/certificate_manager_model.h",
